% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pollination_index.R
\name{calculate_pollination_index}
\alias{calculate_pollination_index}
\title{Calculate pollination index}
\usage{
calculate_pollination_index(prediction_raster, env_rasters, crop_data = NULL)
}
\arguments{
\item{prediction_raster}{SpatRaster}

\item{env_rasters}{SpatRaster}

\item{crop_data}{Optional data.frame with dependence_pollination column}
}
\value{
SpatRaster
}
\description{
Calculate pollination index
}
