% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Clean occurrence coordinates}
\usage{
clean_occurrences(df)
}
\arguments{
\item{df}{data.frame with decimalLongitude, decimalLatitude}
}
\value{
sf object
}
\description{
Clean occurrence coordinates
}
