% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_dependencies}
\alias{crop_dependencies}
\title{Crop pollination dependency data}
\format{
A data frame with four variables:
\describe{
\item{crop}{Character. Common name of the crop.}
\item{dependence_pollination}{Numeric. Illustrative dependency factor from 0 to 1.
0 = no insect pollination dependence, 1 = fully dependent.}
\item{family}{Character. Botanical family.}
\item{type}{Character. Crop category.}
}
}
\source{
Klein, A.M., Vaissiere, B.E., Cane, J.H., Steffan-Dewenter, I.,
Cunningham, S.A., Kremen, C. and Tscharntke, T. (2007). Importance of
pollinators in changing landscapes for world crops.
\emph{Proceedings of the Royal Society B}, 274(1608), 303--313.
\doi{10.1098/rspb.2006.3721}
}
\usage{
crop_dependencies
}
\description{
A dataset of crop species and their dependence on insect pollination,
for package illustration. The stored values have not been independently
verified against the cited literature and are not agronomic prescriptions.
}
\examples{
data(crop_dependencies)
print(crop_dependencies)
barplot(crop_dependencies$dependence_pollination,
        names.arg = crop_dependencies$crop,
        main = "Pollination dependency by crop",
        ylab = "Dependency factor",
        col  = "steelblue")
}
\seealso{
\code{\link{calculate_pollination_index}},
\code{\link{calculate_pollination_deficit}}
}
\keyword{datasets}
