% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recommendations.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations}
\usage{
generate_recommendations(prediction_raster, deficit_raster, landscape)
}
\arguments{
\item{prediction_raster}{SpatRaster, suitability des pollinisateurs}

\item{deficit_raster}{SpatRaster, déficit de pollinisation}

\item{landscape}{list, résultats de analyze_landscape()}
}
\value{
data.frame de recommandations
}
\description{
Génère des recommandations agroécologiques basées sur la suitability,
le déficit de pollinisation et la fragmentation des habitats.
}
