% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_crop_data.R
\name{import_crop_data}
\alias{import_crop_data}
\title{Import crop pollination data}
\usage{
import_crop_data(file_path = NULL)
}
\arguments{
\item{file_path}{Optional path to CSV file}
}
\value{
data.frame with crop dependencies
}
\description{
Import crop pollination data
}
