% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_crop_map.R
\name{import_crop_map}
\alias{import_crop_map}
\title{Import crop map}
\usage{
import_crop_map(path)
}
\arguments{
\item{path}{Chemin vers le fichier vectoriel}
}
\value{
sf object
}
\description{
Lit un fichier vectoriel des cultures (GeoPackage, Shapefile, etc.)
et vérifie la présence de la colonne obligatoire crop_type.
}
\examples{
\dontrun{
  crop_map <- import_crop_map("data-raw/crop_map.gpkg")
}
}
