\name{plot_pollination_deficit}
\alias{plot_pollination_deficit}
\title{Plot illustrative pollination deficit classes}
\usage{plot_pollination_deficit(deficit_raster, title = "Pollination Deficit")}
\arguments{
\item{deficit_raster}{A single-layer terra SpatRaster of deficit values.}
\item{title}{Plot title.}
}
\value{A ggplot object.}
\description{Displays fixed illustrative classes which are not calibrated field risk categories.}
