% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollinator_map.R
\name{plot_pollinator_map}
\alias{plot_pollinator_map}
\title{Plot pollinator suitability map}
\usage{
plot_pollinator_map(r, title = "Pollinator Suitability")
}
\arguments{
\item{r}{SpatRaster}

\item{title}{Character}
}
\value{
ggplot
}
\description{
Plot pollinator suitability map
}
