% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollinatorSDM-package.R
\docType{package}
\name{pollinatorSDM-package}
\alias{pollinatorSDM}
\alias{pollinatorSDM-package}
\title{Package namespace declarations}
\description{
Models pollinator distribution, analyzes suitable habitats, estimates pollination services, and detects pollination deficit zones using species distribution models and landscape analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/idrisslemnouni-crypto/pollinatorSDM}
  \item Report bugs at \url{https://github.com/idrisslemnouni-crypto/pollinatorSDM/issues}
}

}
\author{
\strong{Maintainer}: Idriss Lemnouni \email{251470253+idrisslemnouni-crypto@users.noreply.github.com}

Authors:
\itemize{
  \item Idriss Lemnouni \email{251470253+idrisslemnouni-crypto@users.noreply.github.com}
}

}
\keyword{internal}
