% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{prepare_predictors}
\alias{prepare_predictors}
\title{Prepare environmental predictors}
\usage{
prepare_predictors(occurrences, rasters)
}
\arguments{
\item{occurrences}{data.frame or sf or list}

\item{rasters}{SpatRaster}
}
\value{
data.frame
}
\description{
Prepare environmental predictors
}
