% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train SDM with Random Forest}
\usage{
train_sdm_model(presence, background)
}
\arguments{
\item{presence}{data.frame with predictors}

\item{background}{data.frame with predictors}
}
\value{
randomForest model
}
\description{
Train SDM with Random Forest
}
