% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_landscape.R
\name{analyze_landscape}
\alias{analyze_landscape}
\title{Analyze landscape metrics}
\usage{
analyze_landscape(rasters)
}
\arguments{
\item{rasters}{SpatRaster with landcover layer}
}
\value{
data.frame
}
\description{
Analyze landscape metrics
}
