% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_landscape.R
\name{analyze_landscape}
\alias{analyze_landscape}
\title{Analyze landscape}
\usage{
analyze_landscape(env_rasters)
}
\arguments{
\item{env_rasters}{SpatRaster stack environnemental}
}
\value{
list de métriques paysagères
}
\description{
Analyse la fragmentation des habitats et les métriques paysagères.
}
