% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pollination_deficit.R
\name{calculate_pollination_deficit}
\alias{calculate_pollination_deficit}
\title{Calculate pollination deficit}
\usage{
calculate_pollination_deficit(prediction_raster, pollination_index)
}
\arguments{
\item{prediction_raster}{SpatRaster, offre (suitability)}

\item{pollination_index}{SpatRaster, index calculé}
}
\value{
SpatRaster
}
\description{
Déficit = index de pollinisation - offre (suitability).
Valeurs négatives sont forcées à zéro.
}
