% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pollination_deficit.R
\name{calculate_pollination_deficit}
\alias{calculate_pollination_deficit}
\title{Calculate pollination deficit}
\usage{
calculate_pollination_deficit(pollination_index, crop_needs = 0.5)
}
\arguments{
\item{pollination_index}{SpatRaster}

\item{crop_needs}{SpatRaster or numeric threshold}
}
\value{
SpatRaster
}
\description{
Calculate pollination deficit
}
