% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pollination_index.R
\name{calculate_pollination_index}
\alias{calculate_pollination_index}
\title{Calculate pollination service index}
\usage{
calculate_pollination_index(prediction_raster, crop_demand_raster)
}
\arguments{
\item{prediction_raster}{SpatRaster, suitability des pollinisateurs}

\item{crop_demand_raster}{SpatRaster, demande agricole}
}
\value{
SpatRaster
}
\description{
Multiplie la suitability des pollinisateurs par la demande agricole
(dépendance pollinisation × surface cultivée).
}
