% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pollination_index.R
\name{calculate_pollination_index}
\alias{calculate_pollination_index}
\title{Calculate pollination index}
\usage{
calculate_pollination_index(raster_list)
}
\arguments{
\item{raster_list}{List of SpatRaster (one per species)}
}
\value{
SpatRaster
}
\description{
Calculate pollination index
}
