% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Clean occurrence data}
\usage{
clean_occurrences(data, env_rasters)
}
\arguments{
\item{data}{data.frame avec colonnes decimalLongitude, decimalLatitude, species}

\item{env_rasters}{SpatRaster pour masque terrestre et rarefaction}
}
\value{
list(data = sf object, report = data.frame)
}
\description{
Nettoie les données d'occurrence : supprime doublons, coordonnées nulles,
points marins (hors terre), rarefaction spatiale, et outliers géographiques.
}
