% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Clean occurrence data}
\usage{
clean_occurrences(occurrences)
}
\arguments{
\item{occurrences}{data.frame}
}
\value{
list with cleaned data and report
}
\description{
Clean occurrence data
}
