% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_dependencies.R
\docType{data}
\name{crop_dependencies}
\alias{crop_dependencies}
\title{Crop pollination dependence data}
\format{
data.frame with columns:
\describe{
\item{crop}{Crop name}
\item{dependence_pollination}{Dependence on pollination (0-1)}
\item{family}{Plant family}
\item{type}{Crop type}
}
}
\source{
FAOSTAT and literature synthesis
}
\usage{
crop_dependencies
}
\description{
Crop pollination dependence data
}
\keyword{datasets}
