% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_environmental_layers.R
\name{download_environmental_layers}
\alias{download_environmental_layers}
\title{Download environmental layers}
\usage{
download_environmental_layers()
}
\value{
SpatRaster stack
}
\description{
Télécharge les données climatiques WorldClim et les couches landcover ESA
pour le Maroc via geodata.
}
\examples{
\dontrun{
  env <- download_environmental_layers()
  names(env)
}
}
