% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_environmental_layers.R
\name{download_environmental_layers}
\alias{download_environmental_layers}
\title{Download environmental layers}
\usage{
download_environmental_layers(
  variables = c("temperature", "precipitation", "landcover", "elevation"),
  extent = c(-15, 25, 25, 55)
)
}
\arguments{
\item{variables}{Character vector of variables}

\item{extent}{Numeric extent vector}
}
\value{
SpatRaster
}
\description{
Download environmental layers
}
