% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Evaluate SDM models}
\usage{
evaluate_models(model, test_data)
}
\arguments{
\item{model}{Fitted model}

\item{test_data}{Test dataset}
}
\value{
list with metrics, roc_plot, obs_pred_plot
}
\description{
Evaluate SDM models
}
