% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Evaluate SDM models}
\usage{
evaluate_models(model, test_data)
}
\arguments{
\item{model}{modèle entraîné (randomForest)}

\item{test_data}{data.frame de test avec occurrence et prédicteurs}
}
\value{
list(metrics = data.frame, roc_plot = ggplot)
}
\description{
Calcule AUC, Accuracy, Sensibilité, Spécificité, et génère la courbe ROC.
}
