% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_background_points.R
\name{generate_background_points}
\alias{generate_background_points}
\title{Generate background points}
\usage{
generate_background_points(rasters, n = 1000)
}
\arguments{
\item{rasters}{SpatRaster}

\item{n}{Integer, number of points}
}
\value{
data.frame
}
\description{
Generate background points
}
