% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recommendations.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate management recommendations}
\usage{
generate_recommendations(
  suitability_raster,
  deficit_raster,
  landscape_analysis
)
}
\arguments{
\item{suitability_raster}{SpatRaster de suitability pollinisateurs}

\item{deficit_raster}{SpatRaster du déficit}

\item{landscape_analysis}{list de analyze_landscape()}
}
\value{
data.frame des recommandations
}
\description{
Génère recommandations basées sur déficit de pollinisation et paysage.
}
