% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recommendations.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate management recommendations}
\usage{
generate_recommendations(deficit_class = "eleve")
}
\arguments{
\item{deficit_class}{Character}
}
\value{
Character
}
\description{
Generate management recommendations
}
