% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate HTML report}
\usage{
generate_report(output_file = "pollinator_report.html")
}
\arguments{
\item{output_file}{Character}
}
\value{
Path to report
}
\description{
Generate HTML report
}
