% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_crop_data.R
\name{import_crop_data}
\alias{import_crop_data}
\title{Import crop dependency data}
\usage{
import_crop_data(source = "example")
}
\arguments{
\item{source}{Character source type}
}
\value{
data.frame
}
\description{
Import crop dependency data
}
