% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollination_deficit.R
\name{plot_pollination_deficit}
\alias{plot_pollination_deficit}
\title{Plot pollination deficit map}
\usage{
plot_pollination_deficit(r, title = "Pollination Deficit")
}
\arguments{
\item{r}{SpatRaster}

\item{title}{Character}
}
\value{
ggplot
}
\description{
Plot pollination deficit map
}
