% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollinator_map.R
\name{plot_pollinator_map}
\alias{plot_pollinator_map}
\title{Plot pollinator distribution map}
\usage{
plot_pollinator_map(prediction_raster, title = "Pollinator Suitability")
}
\arguments{
\item{prediction_raster}{SpatRaster}

\item{title}{Character}
}
\value{
ggplot object
}
\description{
Plot pollinator distribution map
}
