% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollinator_occurrences}
\alias{pollinator_occurrences}
\title{Pollinator occurrence example data}
\format{
A data.frame with columns: species, decimalLongitude, decimalLatitude, year
}
\usage{
pollinator_occurrences
}
\description{
Pollinator occurrence example data
}
\keyword{datasets}
