% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pollinator_distribution.R
\name{predict_pollinator_distribution}
\alias{predict_pollinator_distribution}
\title{Predict pollinator distribution on raster}
\usage{
predict_pollinator_distribution(model, rasters)
}
\arguments{
\item{model}{randomForest model}

\item{rasters}{SpatRaster}
}
\value{
SpatRaster
}
\description{
Predict pollinator distribution on raster
}
