% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{prepare_predictors}
\alias{prepare_predictors}
\title{Prepare environmental predictors for SDM}
\usage{
prepare_predictors(occurrences, rasters)
}
\arguments{
\item{occurrences}{sf object}

\item{rasters}{SpatRaster}
}
\value{
data.frame
}
\description{
Prepare environmental predictors for SDM
}
