% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{prepare_predictors}
\alias{prepare_predictors}
\title{Prepare predictors for SDM}
\usage{
prepare_predictors(occurrence_sf, background_sf, env_rasters)
}
\arguments{
\item{occurrence_sf}{sf object d'occurrences (présence)}

\item{background_sf}{sf object de pseudo-absences}

\item{env_rasters}{SpatRaster stack}
}
\value{
data.frame prêt pour SDM
}
\description{
Extrait les valeurs raster, supprime variables corrélées (Pearson > 0.8),
calcule VIF simple, et retourne dataset prêt.
}
