% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_risk_by_region.R
\name{summarize_risk_by_region}
\alias{summarize_risk_by_region}
\title{Summarize pollination deficit by region}
\usage{
summarize_risk_by_region(deficit_raster, regions = NULL)
}
\arguments{
\item{deficit_raster}{SpatRaster}

\item{regions}{sf object}
}
\value{
data.frame
}
\description{
Summarize pollination deficit by region
}
