% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train Species Distribution Model}
\usage{
train_sdm_model(presence_data, background_data, method = "rf")
}
\arguments{
\item{presence_data}{data.frame}

\item{background_data}{data.frame}

\item{method}{Character}
}
\value{
list with model, train_data, test_data
}
\description{
Train Species Distribution Model
}
