% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train SDM model}
\usage{
train_sdm_model(predictor_data, method = "rf", ntree = 500)
}
\arguments{
\item{predictor_data}{data.frame avec variables prédicteurs et occurrence}

\item{method}{character, "rf" uniquement supporté}

\item{ntree}{integer, nombre d'arbres}
}
\value{
modèle randomForest
}
\description{
Entraîne un modèle RandomForest avec présences et background points.
}
